Add files via upload
I modified bind-stats.py and the Zabbix Template to be used with Zabbix-Sender. Because our DNS has more than 200 zones it is impossible use the pooling version.

The bind-stats.py now has the "printcounters" function that output the counters formated to zabbix-sender parser and I send it to txt. After  that I run the "send_bindcounter_zabbix.sh" via crontab or a system.run[] item to send the TXT to zabbix-server (or proxy).